segmentation post-processing into progressing folder
